Made PassGenerator Run Mostly On Another Thread
